Used ternary operator expressio in find_prey
Reported a bug in the algorithm after adding an additional test. Conditional operator was used to correct the fact that
with std::min_element an element is the smallest if no other element compares less than it. This led, if prey appeared after a closer predator in the vector (including predator void itself) to std::min_element returning THAT predator. Problem was effectively detected thanks to the assert that had been placed right after.